requirements.txt
numpy
pandas
matplotlib
opencv-python
jupyter
streamlit
av
pyyaml
Pillow
